pythia160m testing implementation
tested a pythia160m using codesearchnet python subset. the quality of the final model was really bad. it got the general idea of code completion but couldnt generate simple code that was correct.
no pruning vs 20% pruning:
 - -23% accuracy
 - -7% gpu util

which is an impact but nothing of worth. i think my next step is to build an slm compatible with adding and removing capacity seamlessly because i learned that each model is tricky when it comes to pruning as each model is structed differently, whether it's with neurons or strict layers, and whether those connections are linearly or conv